+++ pkg/resource/deploy/step_generator.go
@@ -405,9 +405,14 @@ func (sg *stepGenerator) GenerateSteps(event RegisterResourceEvent) ([]Step, res
 				//       until pulumi/pulumi#624 is resolved, we cannot safely perform this operation on resources
 				//       that have dependent resources (we try to delete the resource while they refer to it).
 				//
-				// The provider is responsible for requesting which of these two modes to use.
-
-				if diff.DeleteBeforeReplace || goal.DeleteBeforeReplace {
+				// The provider is responsible for requesting which of these two modes to use. The user can override
+				// the provider's decision by setting the `deleteBeforeReplace` field of `ResourceOptions` to either
+				// `true` or `false`.
+				deleteBeforeReplace := diff.DeleteBeforeReplace
+				if goal.DeleteBeforeReplace != nil {
+					deleteBeforeReplace = *goal.DeleteBeforeReplace
+				}
+				if deleteBeforeReplace {
 					logging.V(7).Infof("Planner decided to delete-before-replacement for resource '%v'", urn)
 					contract.Assert(sg.plan.depGraph != nil)